move graph checkpoint code out of genesis and into a common location
