+++ iris-mpc-cpu/src/genesis/mod.rs
@@ -1,6 +1,5 @@
 mod batch_generator;
 mod batch_size_config;
-pub mod genesis_checkpoint;
 mod hawk_handle;
 mod hawk_job;
 pub mod plaintext;
@@ -10,7 +9,6 @@ pub mod utils;
 
 pub use batch_generator::{Batch, BatchGenerator, BatchIterator, BatchSize};
 pub use batch_size_config::BatchSizeConfig;
-pub use genesis_checkpoint::PruningMode;
 pub use hawk_handle::Handle;
 pub use hawk_job::{Job, JobRequest, JobResult, SYNC_DONE, SYNC_ERROR, SYNC_RUNNING};
 pub use state_accessor::{
