+++ iris-mpc-cpu/src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod cli;
 pub mod constants;
-pub mod s3_checkpoint;
 pub mod serialization;
